% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-S3.R
\name{as.data.frame.climasus_df}
\alias{as.data.frame.climasus_df}
\title{Coerce climasus_df to data.frame}
\usage{
\method{as.data.frame}{climasus_df}(x, ...)
}
\arguments{
\item{x}{A climasus_df object}

\item{...}{Additional arguments (ignored)}
}
\description{
Coerce climasus_df to data.frame
}
