% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_dlnm.R
\name{coef.climasus_dlnm}
\alias{coef.climasus_dlnm}
\title{Extract coefficients from a climasus_dlnm model}
\usage{
\method{coef}{climasus_dlnm}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_dlnm} object.}

\item{...}{Passed to \code{\link[stats:coef]{stats::coef()}}.}
}
\value{
Named numeric vector of GLM coefficients.
}
\description{
Extract coefficients from a climasus_dlnm model
}
