% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_metaregression.R
\name{coef.climasus_metaregression}
\alias{coef.climasus_metaregression}
\title{Extract pooled coefficients from a climasus_metaregression object}
\usage{
\method{coef}{climasus_metaregression}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_metaregression} object.}

\item{...}{Passed to \code{\link[stats:coef]{stats::coef()}}.}
}
\value{
Named numeric vector of meta-regression coefficients.
}
\description{
Extract pooled coefficients from a climasus_metaregression object
}
