% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_pool.R
\name{coef.climasus_pool}
\alias{coef.climasus_pool}
\title{Extract pooled coefficients from a climasus_pool object}
\usage{
\method{coef}{climasus_pool}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_pool} object.}

\item{...}{Passed to \code{\link[stats:coef]{stats::coef()}}.}
}
\value{
Named numeric vector of pooled cross-basis coefficients.
}
\description{
Extract pooled coefficients from a climasus_pool object
}
