% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_climate_compute_coldwaves.R
\name{cw_get_events}
\alias{cw_get_events}
\title{Extract coldwave events table from a climasus_cw result}
\usage{
cw_get_events(cw_result, method_filter = NULL)
}
\arguments{
\item{cw_result}{A \code{climasus_cw} list returned by
\code{sus_climate_compute_coldwaves()}.}

\item{method_filter}{Character vector. If supplied, keep only these methods.}
}
\value{
A tibble of coldwave events.
}
\description{
Extract coldwave events table from a climasus_cw result
}
