% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_create_variables.R
\name{detect_date_column}
\alias{detect_date_column}
\title{Detect Date Column}
\usage{
detect_date_column(df, system)
}
\arguments{
\item{df}{Data frame}
}
\value{
Name of date column
}
\description{
Auto-detects the date column in a data frame based on common patterns.
}
\keyword{internal}
