% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_data_filter_cid.R
\name{filter_by_icd_range}
\alias{filter_by_icd_range}
\title{Filter by ICD-10 code range}
\usage{
filter_by_icd_range(df, icd_column, ranges)
}
\arguments{
\item{df}{Data frame}

\item{icd_column}{Name of ICD column}

\item{ranges}{Character vector of ranges}
}
\value{
Filtered data frame
}
\description{
Internal function to filter data by ICD-10 code ranges.
}
\keyword{internal}
