% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_create_variables.R
\name{generate_age_labels}
\alias{generate_age_labels}
\title{Generate Age Labels}
\usage{
generate_age_labels(breaks, lang)
}
\arguments{
\item{breaks}{Numeric vector of age breaks}

\item{lang}{Language for labels}
}
\value{
Character vector of labels
}
\description{
Generates age group labels from breaks.
}
\keyword{internal}
