% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_data_aggregate.R
\name{get_brazilian_season_start}
\alias{get_brazilian_season_start}
\title{Get Brazilian Season Start Date}
\usage{
get_brazilian_season_start(dates)
}
\arguments{
\item{dates}{A vector of Date objects}
}
\value{
A vector of Date objects representing the start of each season
}
\description{
Returns the first day of the Brazilian season for a given date.
Seasons are defined as:
\itemize{
\item Summer (Verao): Dec-Jan-Feb (starts Dec 1st)
\item Autumn (Outono): Mar-Apr-May (starts Mar 1st)
\item Winter (Inverno): Jun-Jul-Aug (starts Jun 1st)
\item Spring (Primavera): Sep-Oct-Nov (starts Sep 1st)
}
}
\examples{
\dontrun{
dates <- as.Date(c("2023-01-15", "2023-06-20", "2023-12-25"))
get_brazilian_season_start(dates)
# Returns: "2022-12-01" "2023-06-01" "2023-12-01"
}
}
\keyword{internal}
