% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_climasus_cache_info}
\alias{get_climasus_cache_info}
\title{Get cache information}
\usage{
get_climasus_cache_info(cache_dir = "~/.climasus4r_cache", verbose = TRUE)
}
\arguments{
\item{cache_dir}{Character. Cache directory to inspect. Default is "~/.climasus4r_cache".}

\item{verbose}{Logical. If TRUE, prints detailed information.}
}
\value{
A list with cache statistics.
}
\description{
Get cache information
}
\examples{
\dontrun{
cache_info <- get_climasus_cache_info()
}
}
