% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_create_variables.R
\name{get_day_names}
\alias{get_day_names}
\title{Get Day Names}
\usage{
get_day_names(day_num, lang)
}
\arguments{
\item{day_num}{Numeric vector of day numbers (1-7, Monday=1)}

\item{lang}{Language code}
}
\value{
Character vector of day names
}
\description{
Returns day of week names in the specified language.
}
\keyword{internal}
