% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_create_variables.R
\name{get_month_names}
\alias{get_month_names}
\title{Get Month Names}
\usage{
get_month_names(month_num, lang)
}
\arguments{
\item{month_num}{Numeric vector of month numbers (1-12)}

\item{lang}{Language code}
}
\value{
Character vector of month names
}
\description{
Returns month names in the specified language.
}
\keyword{internal}
