% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_create_variables.R
\name{get_season}
\alias{get_season}
\title{Get Season}
\usage{
get_season(month_num, hemisphere, lang)
}
\arguments{
\item{month_num}{Numeric vector of month numbers (1-12)}

\item{hemisphere}{Hemisphere ("south" or "north")}

\item{lang}{Language code}
}
\value{
Character vector of season names
}
\description{
Returns season based on month and hemisphere.
}
\keyword{internal}
