% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_data_standardize.R
\name{get_ui_messages_standardize}
\alias{get_ui_messages_standardize}
\title{Get UI Messages}
\usage{
get_ui_messages_standardize(lang = "en")
}
\arguments{
\item{lang}{Character. Language code: "en", "pt", "es"}
}
\value{
List of translated UI messages
}
\description{
Returns user interface messages in the specified language
}
\keyword{internal}
