% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_af.R
\name{print.climasus_af}
\alias{print.climasus_af}
\title{Print a climasus_af object}
\usage{
\method{print}{climasus_af}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_af} object.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_af object
}
