% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_burden.R
\name{print.climasus_burden}
\alias{print.climasus_burden}
\title{Print a climasus_burden object}
\usage{
\method{print}{climasus_burden}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_burden} object from \code{\link[=sus_mod_burden]{sus_mod_burden()}}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_burden object
}
