% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_casecrossover.R
\name{print.climasus_casecrossover}
\alias{print.climasus_casecrossover}
\title{Print a climasus_casecrossover object}
\usage{
\method{print}{climasus_casecrossover}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_casecrossover} object from \code{sus_mod_casecrossover()}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_casecrossover object
}
