% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-S3.R
\name{print.climasus_df}
\alias{print.climasus_df}
\title{Print method for climasus_df}
\usage{
\method{print}{climasus_df}(x, n = 10, ...)
}
\arguments{
\item{x}{A climasus_df object}

\item{n}{Integer. Number of rows to print}

\item{...}{Additional arguments passed to print.data.frame}
}
\description{
Print method for climasus_df
}
