% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_dlnm.R
\name{print.climasus_dlnm}
\alias{print.climasus_dlnm}
\title{Print a climasus_dlnm model}
\usage{
\method{print}{climasus_dlnm}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_dlnm} object from \code{sus_mod_dlnm()}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_dlnm model
}
