% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_excess.R
\name{print.climasus_excess}
\alias{print.climasus_excess}
\title{Print a climasus_excess object}
\usage{
\method{print}{climasus_excess}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_excess} object from \code{sus_mod_excess()}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_excess object
}
