% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_metaregression.R
\name{print.climasus_metaregression}
\alias{print.climasus_metaregression}
\title{Print a climasus_metaregression object}
\usage{
\method{print}{climasus_metaregression}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_metaregression} object from \code{sus_mod_metaregression()}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_metaregression object
}
