% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_pool.R
\name{print.climasus_pool}
\alias{print.climasus_pool}
\title{Print a climasus_pool object}
\usage{
\method{print}{climasus_pool}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_pool} object from \code{sus_mod_pool()}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_pool object
}
