% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_data_quality_report.R
\name{print.climasus_quality_report}
\alias{print.climasus_quality_report}
\title{Print a climasus4r data quality report}
\usage{
\method{print}{climasus_quality_report}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_quality_report} object from \code{\link[=sus_data_quality_report]{sus_data_quality_report()}}.}

\item{...}{Unused.}
}
\description{
Print a climasus4r data quality report
}
\keyword{internal}
