% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_sensitivity.R
\name{print.climasus_sensitivity}
\alias{print.climasus_sensitivity}
\title{Print a climasus_sensitivity object}
\usage{
\method{print}{climasus_sensitivity}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_sensitivity} object from \code{sus_mod_sensitivity()}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_sensitivity object
}
