% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_swot.R
\name{print.climasus_swot}
\alias{print.climasus_swot}
\title{Print a climasus_swot object}
\usage{
\method{print}{climasus_swot}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_swot} object from \code{\link[=sus_mod_swot]{sus_mod_swot()}}.}

\item{...}{Ignored.}
}
\description{
Print a climasus_swot object
}
