% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_vulnerability_index.R
\name{print.climasus_vi}
\alias{print.climasus_vi}
\title{Print a climasus_vi object}
\usage{
\method{print}{climasus_vi}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_vi} object from \code{\link[=sus_mod_vulnerability_index]{sus_mod_vulnerability_index()}}.}

\item{...}{Unused.}
}
\value{
\code{x} invisibly.
}
\description{
Print a climasus_vi object
}
