% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-S3.R
\name{[.climasus_df}
\alias{[.climasus_df}
\title{Subsetting method for climasus_df}
\usage{
\method{[}{climasus_df}(x, i, j, drop = FALSE)
}
\arguments{
\item{x}{A climasus_df object}

\item{i}{Row indices}

\item{j}{Column indices}

\item{drop}{Logical}
}
\description{
Subsetting method for climasus_df
}
