% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-S3.R
\name{[[.climasus_df}
\alias{[[.climasus_df}
\title{Column extraction for climasus_df}
\usage{
\method{[[}{climasus_df}(x, i)
}
\arguments{
\item{x}{A climasus_df object}

\item{i}{Column index or name}
}
\description{
Column extraction for climasus_df
}
