% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_burden.R
\name{summary.climasus_burden}
\alias{summary.climasus_burden}
\title{Summarise a climasus_burden object}
\usage{
\method{summary}{climasus_burden}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_burden} object from \code{\link[=sus_mod_burden]{sus_mod_burden()}}.}

\item{...}{Unused.}
}
\value{
\code{object} invisibly.
}
\description{
Summarise a climasus_burden object
}
