% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_casecrossover.R
\name{summary.climasus_casecrossover}
\alias{summary.climasus_casecrossover}
\title{Summarise a climasus_casecrossover object}
\usage{
\method{summary}{climasus_casecrossover}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_casecrossover} object from \code{sus_mod_casecrossover()}.}

\item{...}{Unused.}
}
\value{
\code{object} invisibly.
}
\description{
Summarise a climasus_casecrossover object
}
