% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_excess.R
\name{summary.climasus_excess}
\alias{summary.climasus_excess}
\title{Summarise a climasus_excess object}
\usage{
\method{summary}{climasus_excess}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_excess} object from \code{sus_mod_excess()}.}

\item{...}{Unused.}
}
\value{
\code{object} invisibly.
}
\description{
Summarise a climasus_excess object
}
