% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_ml.R
\name{summary.climasus_ml}
\alias{summary.climasus_ml}
\title{Summarise a climasus_ml object}
\usage{
\method{summary}{climasus_ml}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_ml} object from \code{\link[=sus_mod_ml]{sus_mod_ml()}}.}

\item{...}{Unused.}
}
\value{
\code{object} invisibly.
}
\description{
Summarise a climasus_ml object
}
