% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_pool.R
\name{summary.climasus_pool}
\alias{summary.climasus_pool}
\title{Summarise a climasus_pool object}
\usage{
\method{summary}{climasus_pool}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_pool} object from \code{sus_mod_pool()}.}

\item{...}{Unused.}
}
\value{
\code{object} invisibly.
}
\description{
Summarise a climasus_pool object
}
