% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_sensitivity.R
\name{summary.climasus_sensitivity}
\alias{summary.climasus_sensitivity}
\title{Summarise a climasus_sensitivity object}
\usage{
\method{summary}{climasus_sensitivity}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_sensitivity} object from \code{sus_mod_sensitivity()}.}

\item{...}{Unused.}
}
\value{
\code{object} invisibly.
}
\description{
Summarise a climasus_sensitivity object
}
