% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_spacetime_predict.R
\name{summary.climasus_spacetime_pred}
\alias{summary.climasus_spacetime_pred}
\title{Summary method for climasus_spacetime_pred objects}
\usage{
\method{summary}{climasus_spacetime_pred}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_spacetime_pred} object.}

\item{...}{Additional arguments passed to \code{\link[=print.climasus_spacetime_pred]{print.climasus_spacetime_pred()}}.}
}
\value{
Invisibly returns \code{object}.
}
\description{
Alias for \code{\link[=print.climasus_spacetime_pred]{print.climasus_spacetime_pred()}}.
}
