% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_swot.R
\name{summary.climasus_swot}
\alias{summary.climasus_swot}
\title{Summarise a climasus_swot object}
\usage{
\method{summary}{climasus_swot}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_swot} object from \code{\link[=sus_mod_swot]{sus_mod_swot()}}.}

\item{...}{Ignored.}
}
\description{
Summarise a climasus_swot object
}
