% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_data_ts_quality.R
\name{summary.climasus_ts_quality}
\alias{summary.climasus_ts_quality}
\title{Summarise a climasus_ts_quality object}
\usage{
\method{summary}{climasus_ts_quality}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_ts_quality} object.}

\item{...}{Unused.}
}
\value{
\code{object} invisibly.
}
\description{
Summarise a climasus_ts_quality object
}
