% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_dlnm.R
\name{tidy.climasus_dlnm}
\alias{tidy.climasus_dlnm}
\title{Reduce climasus_dlnm to a tidy one-row tibble}
\usage{
\method{tidy}{climasus_dlnm}(x, ...)
}
\arguments{
\item{x}{A \code{climasus_dlnm} object.}

\item{...}{Unused.}
}
\value{
A one-row tibble with key model statistics.
}
\description{
Convenience extractor compatible with \code{dplyr::bind_rows()} for pooling
results across multiple model runs.
}
