% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_create_variables.R
\name{try_detect_age_column}
\alias{try_detect_age_column}
\title{Try to Detect Existing Age Column}
\usage{
try_detect_age_column(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
Name of age column or NULL if not found
}
\description{
Try to Detect Existing Age Column
}
\keyword{internal}
