% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils}
\alias{utils}
\title{Multilingual Translation System for CLIMASUS4r}
\description{
Internal functions to support Portuguese, Spanish, and English translations
throughout the package.
}
\keyword{internal}
