% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sus_mod_metaregression.R
\name{vcov.climasus_metaregression}
\alias{vcov.climasus_metaregression}
\title{Extract variance-covariance from a climasus_metaregression object}
\usage{
\method{vcov}{climasus_metaregression}(object, ...)
}
\arguments{
\item{object}{A \code{climasus_metaregression} object.}

\item{...}{Passed to \code{\link[stats:vcov]{stats::vcov()}}.}
}
\value{
Variance-covariance matrix of meta-regression coefficients.
}
\description{
Extract variance-covariance from a climasus_metaregression object
}
